# docs/swagger.yaml
definitions:
...
  CommandResponse:
    properties:
      data:
        description: Data to pass additional data
        type: object
      id:
        description: ID the id of the related entity
        type: string
      reason:
        description: Reason exist when command is rejected
        type: string
      status:
        description: Status can be accepted or rejected
        type: string
      type:
        description: Type the type of the related entity
        type: string
    type: object
  CreateTaskResponse:
    properties:
      description:
        type: string
      list:
        type: string
      name:
        type: string
      project:
        type: string
    required:
    - name
    type: object
...
  TriggerRunRequestBody:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  task.Metadata:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
  task.Spec:
    properties:
      dueDate:
        type: string
    type: object
  task.Status:
    properties:
      completed:
        type: boolean
    type: object
  task.Task:
    properties:
      id:
        type: string
      metadata:
        $ref: '#/definitions/task.Metadata'
      spec:
        $ref: '#/definitions/task.Spec'
      status:
        $ref: '#/definitions/task.Status'
      type:
        type: string
    required:
    - metadata
    - spec
    type: object
  task.completeReopenTaskRequestBody:
    properties:
      task:
        type: string
    type: object
  task.deleteTaskRequestBody:
    properties:
      id:
        type: string
    type: object
...
host: localhost
info:
  contact:
    name: Oleg Sucharevich
  description: Peteq OpenAPI spec.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Peteq API
  version: "1.0"
paths:
...
  /c/task/complete:
    post:
      consumes:
      - application/json
      description: Complete task
      parameters:
      - description: Complete Task Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.completeReopenTaskRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Task Command API
  /c/task/create:
    post:
      consumes:
      - application/json
      description: Create new task
      parameters:
      - description: Create Task Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateTaskResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Task Command API
  /c/task/delete:
    post:
      consumes:
      - application/json
      description: Delete task
      parameters:
      - description: Delete Task Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.deleteTaskRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Task Command API
  /c/task/reopen:
    post:
      consumes:
      - application/json
      description: Reopen task
      parameters:
      - description: Reopen Task Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.completeReopenTaskRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Task Command API
  /c/task/update:
    post:
      consumes:
      - application/json
      description: Update task
      parameters:
      - description: Update Task Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Task Command API
  /c/trigger/run:
    post:
      consumes:
      - application/json
      description: Trigger run
      parameters:
      - description: Trigger run
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/TriggerRunRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommandResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - Trigger Command API
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"